+++ src/main/kotlin/com/github/copyinaction/SmarterStoreApiApplication.kt
@@ -4,7 +4,10 @@ import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing
 
+import org.springframework.scheduling.annotation.EnableScheduling
+
 @EnableJpaAuditing
+@EnableScheduling // 스케줄링 활성화
 @SpringBootApplication
 class SmarterStoreApiApplication